+++ core/src/com/bisoft12/cenforpg/StartGame.java
@@ -13,11 +13,13 @@ public class StartGame extends Game {
 		//Resources.MAIN.setScreen(new LoadScreen());
 		//Resources.MAIN.setScreen(new CityScreen());
 		//Resources.MAIN.setScreen(new DungeonScreen());
-	  Resources.MAIN.setScreen(new HouseScreen());
-		//Resources.MAIN.setScreen(new MerchantScreen());
+	  //Resources.MAIN.setScreen(new HouseScreen());
+		Resources.MAIN.setScreen(new MerchantScreen());
 		//Resources.MAIN.setScreen(new TerrainMonster());
 		//Resources.MAIN.setScreen(new FightScreen());
-		//Resources.MAIN.setScreen(new MenuScreen());
+		//Resources.MAIN.setScreen(new CreateCharacterScreen());
+		//Resources.MAIN.setScreen(new FightScreen());
+		//Resources.MAIN.setScreen(new LoreScreen());
 
 	}
