+++ core/src/com/bisoft12/cenforpg/patterns/Comportamiento/Patron_Estado/concreto/Recuperar.java
@@ -1,29 +1,23 @@
 package com.bisoft12.cenforpg.patterns.Comportamiento.Patron_Estado.concreto;
 import com.badlogic.gdx.utils.Timer;
+import com.bisoft12.cenforpg.patterns.Comportamiento.Main.Gestor_Estado;
 import com.bisoft12.cenforpg.patterns.Comportamiento.Patron_Estado.abstracto.Estado;
 import com.bisoft12.cenforpg.patterns.Comportamiento.Patron_Estado.objeto.Casa;
-
-import java.sql.Time;
+import static java.lang.System.out;
 
 public class Recuperar extends Estado {
+    private Gestor_Estado gEstado = new Gestor_Estado();
 
-    private Casa casa;
     @Override
     public String activarEstado() {
-        return "Recuperando vidas. \n" + detenerJuego();
+        for (int contk = 10; contk >= 0; contk--) {
+            out.println("Recuperando vidas");
+        }
+            out.println("Tiempo de Pelear!");
+            gEstado.Cambiar_Estado(2);
+        return "Pelear";
     }
-    private String detenerJuego() {
-        this.casa = casa;
 
-        Timer timer = new Timer();
-        Timer.schedule(new Timer.Task() {
-            @Override
-            public void run() {
-                Recuperar.this.casa.setEstado(new Pelear());
-            }
-        }, 5000);
 
-        return ("Necesitas recuperarte antes de pelear");
-    }
 
 }//
