--- core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/MetodoFab_Pocion.java
@@ -1,6 +1,0 @@
-package com.bisoft12.cenforpg.patterns.Creational.MetodoFab;
-
-public interface MetodoFab_Pocion {
-    public Pocion crearPocion (int pTipo);
-
-}//
