--- core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/Pocion_LanzaLlamas.java
@@ -1,12 +1,0 @@
-package com.bisoft12.cenforpg.patterns.Creational.MetodoFab;
-
-public class Pocion_LanzaLlamas extends Pocion {
-
-    // ID = 1
-    public Pocion_LanzaLlamas(){
-         setNombre("Lanza Llamas");
-    }
-
-
-
-}//
