--- core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/Pocion_Recuperativa.java
@@ -1,16 +1,0 @@
-package com.bisoft12.cenforpg.patterns.Creational.MetodoFab;
-
-public class Pocion_Recuperativa extends Pocion{
-
-    // ID = 3
- public Pocion_Recuperativa() {
-
-     setNombre("Mana");
- }
-
-    /****************** SECCION: Implementación de las funciones abstractas *******************/
-
-
-
-
-}//
