--- core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/Pocion_SoploHielo.java
@@ -1,11 +1,0 @@
-package com.bisoft12.cenforpg.patterns.Creational.MetodoFab;
-
-public class Pocion_SoploHielo extends Pocion {
-    // ID = 2
-    public Pocion_SoploHielo() {
-        setNombre("Soplo de Hielo");
-    }
-
-
-
-}//
