--- core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/Pocion_VitalidadGrupal.java
@@ -1,15 +1,0 @@
-package com.bisoft12.cenforpg.patterns.Creational.MetodoFab;
-
-public class Pocion_VitalidadGrupal extends Pocion{
-
-    // ID = 4
- public Pocion_VitalidadGrupal() {
-    setNombre("Mana Colectivo");
- }
-
-    /****************** SECCION: Implementación de las funciones abstractas *******************/
-
-
-
-
-}//
