+++ core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/creador/MetodoFab_Pocion.java
@@ -1,0 +1,8 @@
+package com.bisoft12.cenforpg.patterns.Creational.MetodoFab.creador;
+
+import com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto.Pocion;
+
+public interface MetodoFab_Pocion {
+    public Pocion crearPocion (int pTipo);
+
+}//
