+++ core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/producto_concreto/Pocion_LanzaLlamas.java
@@ -1,0 +1,14 @@
+package com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto_concreto;
+
+import com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto.Pocion;
+
+public class Pocion_LanzaLlamas extends Pocion {
+
+    // ID = 1
+    public Pocion_LanzaLlamas(){
+         setNombre("Lanza Llamas");
+    }
+
+
+
+}//
