+++ core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/producto_concreto/Pocion_Recuperativa.java
@@ -1,0 +1,18 @@
+package com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto_concreto;
+
+import com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto.Pocion;
+
+public class Pocion_Recuperativa extends Pocion {
+
+    // ID = 3
+ public Pocion_Recuperativa() {
+
+     setNombre("Mana");
+ }
+
+    /****************** SECCION: Implementación de las funciones abstractas *******************/
+
+
+
+
+}//
