+++ core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/producto_concreto/Pocion_SoploHielo.java
@@ -1,0 +1,13 @@
+package com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto_concreto;
+
+import com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto.Pocion;
+
+public class Pocion_SoploHielo extends Pocion {
+    // ID = 2
+    public Pocion_SoploHielo() {
+        setNombre("Soplo de Hielo");
+    }
+
+
+
+}//
