+++ core/src/com/bisoft12/cenforpg/patterns/Creational/MetodoFab/producto_concreto/Pocion_VitalidadGrupal.java
@@ -1,0 +1,17 @@
+package com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto_concreto;
+
+import com.bisoft12.cenforpg.patterns.Creational.MetodoFab.producto.Pocion;
+
+public class Pocion_VitalidadGrupal extends Pocion {
+
+    // ID = 4
+ public Pocion_VitalidadGrupal() {
+    setNombre("Mana Colectivo");
+ }
+
+    /****************** SECCION: Implementación de las funciones abstractas *******************/
+
+
+
+
+}//
