+++ core/src/com/bisoft12/cenforpg/screen/MerchantScreen.java
@@ -3,28 +3,16 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.graphics.g2d.BitmapFont;
-import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
-import com.badlogic.gdx.scenes.scene2d.Stage;
-import com.badlogic.gdx.scenes.scene2d.ui.Skin;
-import com.badlogic.gdx.scenes.scene2d.ui.Table;
-import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
-import com.badlogic.gdx.utils.viewport.FitViewport;
-import com.bisoft12.cenforpg.characters.Player;
 import com.bisoft12.cenforpg.elements.Images;
 import com.bisoft12.cenforpg.elements.Text;
 import com.bisoft12.cenforpg.io.Inputs;
 import com.bisoft12.cenforpg.patterns.Creational.FabricaAbstracta.Gestor.FabricaCharacter;
-import com.bisoft12.cenforpg.patterns.Creational.MetodoFab.Gestor_Pocion;
 import com.bisoft12.cenforpg.screen.MerchantMenu.Merchant_MenuArmas;
 import com.bisoft12.cenforpg.screen.MerchantMenu.Merchant_MenuPociones;
-import com.bisoft12.cenforpg.utils.Pantalla;
 import com.bisoft12.cenforpg.utils.Render;
 import com.bisoft12.cenforpg.utils.Resources;
 
-import java.awt.*;
 import java.util.ArrayList;
 
 public class MerchantScreen implements Screen {
@@ -51,7 +39,7 @@ public class MerchantScreen implements Screen {
     public void show() {
         generateMenu();
         Gdx.input.setInputProcessor(this.input);
-        sound = Gdx.audio.newSound(Gdx.files.internal("music/MerchantMusic.mp3"));
+        sound = Gdx.audio.newSound(Gdx.files.internal("music/BackgroundMusic.mp3"));
         sound.play();
     }
 
@@ -111,19 +99,21 @@ public class MerchantScreen implements Screen {
                 this.dispose();
                 sound.stop();
                 break;
+
             case 1: //Armaduras
                 //Enviar al Patron Prototipo el ID
                 Resources.MAIN.setScreen(new TerrainMonster());
                 this.dispose();
                 sound.stop();
                 break;
+
             case 2: //Pociones
                 //gCharacter.get
-
                 Resources.MAIN.setScreen(new Merchant_MenuPociones());
                 this.dispose();
                 sound.stop();
                 break;
+
             case 3: //Salir
                 Resources.MAIN.setScreen(new CityScreen());
                 this.dispose();
